API_ProtoRH/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
sqlalchemy-utils
python-dotenv
pyjwt
python-multipart